Added MacOSX request for Microphone usage.
git-svn-id: https://svn.code.sf.net/p/beebeep/code@1342 127b15f6-8070-407b-945c-913c06b37495

+++ src/utils/BeeUtils.h
@@ -94,7 +94,7 @@ namespace Bee
   QString removeInvalidCharactersForFilePath( const QString& );
 
   QString bytesToString( FileSizeType, int precision = -1 );
-  QString timeToString( int msec );
+  QString timeToString( qint64 msec );
   qint64 roundFromDouble( double d );
   qint64 bytesPerSecond( FileSizeType transferred_byte_size, int time_elapsed_ms );
   QString transferTimeLeft( FileSizeType bytes_transferred, FileSizeType total_bytes, FileSizeType starting_position, int elapsed_time );
